Top Hat bugs fixed

+++ week2/M1_Team8_TopHat.m
@@ -2,4 +2,4 @@ function OutputImg = M1_Team8_TopHat (InputImg, structel, struCent, Times)
 
 OpenedImg = M1_Team8_Opening(InputImg,structel,struCent,Times);
 
-OutputImg = InputImg - OpenedImg;
+OutputImg = double(InputImg) - OpenedImg;
